server/requirements.txt
django
djangorestframework
django-cors-headers
pygments
django-q
redis
python-dotenv
psycopg2-binary
croniter
discord-webhook
